Fix merge conflict (#3769)
